Change udp parsing to use 'read' instead of 'readBytes'
Verified the parsing of udp packets when using visualizer patterns was
taking ~ 1000ms, resulting in a very low fps when using these particular
patterns. Narrowed the timing down to Udp.readBytes(). After being
unable to find a solid reference for readBytes or why it might be taking
so long to return, I tested with a simple 'read()' call and am now
getting normal speeds from the visualizer patterns.

issue #204